voice agent/requirements.txt
flask
python-dotenv
google-generativeai
waitress
supabase
psycopg[binary]
requests
gunicorn


